+++ client/src/lib/functions.ts
@@ -338,3 +338,5 @@ export function dependFieldReplacer<
 
 export const btoa = (s: string) => Buffer.from(s).toString("base64")
 export const atob = (s: string) => Buffer.from(s, "base64").toString()
+export const NSSLASH = "\u2215" // used to be hardcoded as "／", aka "\uFF0F"
+export const convertSlash = (x: string) => x && x.replace(/\//g, NSSLASH)