+++ client/src/scm/abapGit/documentProvider.ts
@@ -9,7 +9,10 @@ class GitDocProvider implements TextDocumentContentProvider {
   async provideTextDocumentContent(uri: Uri) {
     if (uri.scheme !== GITSCHEME)
       throw new Error(`Unexpected URI scheme ${uri.scheme}`)
-    const { key = "", path = "" } = JSON.parse(atob(uri.query))
+    const { key = "", path = "" } = JSON.parse(atob(uri.query)) as {
+      path: string
+      key: string
+    }
 
     const gitScm = scmData(scmKey(uri.authority, key))
     if (!key || !path || !gitScm) throw new Error(`Invalid URL`)
@@ -17,7 +20,11 @@ class GitDocProvider implements TextDocumentContentProvider {
     if (!server) throw new Error(`No active connection for ${uri.authority}`)
     // by now I can take for granted that login happened
     const { user, password } = gitScm.credentials || {}
-    return server.client.getObjectSource(path, user, password)
+    return server.client.getObjectSource(
+      path.replace(/#/g, "%23"),
+      user,
+      password
+    )
   }
 }
 
@@ -25,7 +32,5 @@ workspace.registerTextDocumentContentProvider(GITSCHEME, new GitDocProvider())
 
 export const gitUrl = (data: ScmData, path: string, file: GitStagingFile) => {
   const query = btoa(JSON.stringify({ key: data.repo.key, path }))
-  return Uri.parse(
-    `${GITSCHEME}://${data.connId}${file.path}${file.name}?${query}`
-  )
+  return Uri.parse(`${GITSCHEME}://${data.connId}?${query}`)
 }